pyproject.toml
[project]
name = "wt"
version = "0.1.0"
description = "Simple tool to work with git worktrees"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.44",
    "iterfzf>=1.6.0.60.3",
    "questionary>=2.1.0",
]
